source/api/requirements.txt
adal==1.2.2
antlr4-python3-runtime==4.8
appdirs==1.4.4
arrow==1.2.1
asgiref==3.4.1
attrs==21.2.0
boto3==1.17.20
botocore==1.20.112
cattrs==1.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
colorlog==4.1.0
cpe==1.2.1
cryptography==35.0.0
cybox==2.1.0.21
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
Flask==1.1.2
flatten-json==0.1.7
fqdn==1.5.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
jsonmerge==1.7.0
jsonpointer==2.2
jsonschema==4.0.0
lxml==4.6.4
maec==4.1.0.17
MarkupSafe==2.0.1
mixbox==1.0.5
mysql-connector-python==8.0.25
netaddr==0.8.0
ordered-set==4.0.2
Pillow==9.1.1
pluralizer==1.1.0
protobuf==3.19.1
psycopg2==2.9.2
pycountry==20.7.3
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==20.0.1
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-cache==0.7.5
requests-toolbelt==0.9.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
s3transfer==0.3.7
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
stix==1.2.0.11
stix-pattern-translator==0.2.0
stix-shifter==3.6.1
stix-shifter-modules-mysql==3.6.1
stix-shifter-modules-qradar==3.6.1
stix-shifter-utils==3.6.1
stix2==3.0.1
stix2-matcher==2.0.1
stix2-patterns==1.3.0
stix2-validator==1.1.2
stixmarx==1.0.8
uri-template==1.1.0
url-normalize==1.4.3
urllib3==1.26.7
weakrefmethod==1.0.3
webcolors==1.11.1
Werkzeug==2.0.2
xmltodict==0.12.0
